+++ mlem_method/simulated_representations.py
@@ -43,7 +43,7 @@ class SimulatedRepresentations(BaseModel):
         import pandas as pd
         from sklearn.datasets import make_sparse_spd_matrix, make_spd_matrix
 
-        features = self.dataset.features
+        features = self.dataset.coordinates
         n_features = len(features)
         if self.sparse_spd:
             W = make_sparse_spd_matrix(
@@ -55,19 +55,19 @@ class SimulatedRepresentations(BaseModel):
             W = make_spd_matrix(n_features, random_state=seed_from_basemodel(self))
         W /= np.linalg.norm(W, ord="fro")
         self._W = W
-        W_triu = W[*self.dataset.triu_indices]
-        self._gt_weights = pd.DataFrame(
-            {"Feature": self.dataset.pfeatures, "GTWeight": W_triu}
-        )
+        W_triu = W[*np.triu_indices(n_features)]
+        self._gt_weights = pd.DataFrame({"Feature": self.dataset.pcoordinates, "GTWeight": W_triu})
 
     def __call__(self):
         if self._repr is None:
             import torch
             from sklearn.manifold import MDS
 
-            X = self.dataset.encode()
+            X = self.dataset.encode()[0]
 
-            pX = (X[:, None] - X).abs().clip(0, 1)
+            pX = X[:, None] - X
+            if not self.dataset.mahalanobis:
+                pX = pX.abs().clip(0, 1)
             gt_dist = (pX @ self.W * pX).sum(dim=2)
             mds = MDS(
                 n_components=self.n_components,